tests

+++ ext/EvoTreesCUDAExt/init.jl
@@ -67,7 +67,7 @@ function EvoTrees.init_core(params::EvoTrees.EvoTypes, ::Type{<:EvoTrees.GPU}, d
     !isnothing(offset) && (pred .+= CuArray(offset'))
 
     # initialize gradients
-    h∇_cpu = zeros(Float64, 2 * K + 1, maximum(featbins), length(featbins))
+    h∇_cpu = zeros(Float32, 2 * K + 1, maximum(featbins), length(featbins))
     h∇ = CuArray(h∇_cpu)
     ∇ = CUDA.zeros(T, 2 * K + 1, nobs)
     @assert (length(y) == length(w) && minimum(w) > 0)
